+++ Cargo.toml
@@ -16,7 +16,8 @@ indicatif = "0.18.0"
 reqwest = { version = "0.12.23", features = ["rustls-tls"] }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
-tokio = { version = "1.47.1", features = ["macros", "rt-multi-thread"] }
+tokio = { version = "1.47.1", features = ["macros", "rt-multi-thread", "fs"] }
+toml = "0.9.8"
 
 # The profile that 'dist' will build with
 [profile.dist]
