+++ mise.toml
@@ -5,13 +5,12 @@
 "github:rojo-rbx/rojo" = "7"
 stylua = "latest"
 
-[tasks.build]
+[tasks.build-plugin]
 run = "rojo build -o plugin.rbxm"
 
-[tasks.insert]
-run = "drillbit"
-depends = ["build"]
+[tasks.install-cli]
+run = "cargo install --locked --path ."
 
-[tasks.test]
-run = 'cargo run -- -p ReplicatedStorage/shared,ServerScriptService --testMatch "**/*.spec"'
-depends = ["insert"]
+[tasks.dev]
+run = "drillbit"
+depends = ["build-plugin", "install-cli"]
