+++ plugin/patch/CoreScriptSyncService.luau
@@ -5,8 +5,7 @@ local getRelativePath = require((script.Parent :: any).getRelativePath)
 local CoreScriptSyncService = {}
 
 function CoreScriptSyncService:GetScriptFilePath(path)
-	print(`CoreScriptSyncService:GetScriptFilePath({path})`)
-	return getRelativePath(path)
+	return getRelativePath(path, game:FindFirstChild("jest-companion", true))
 end
 
 return CoreScriptSyncService
